BREAKINGPOINT: Game renders but stays in a loop, prevents input
STATUS

The following is a "screeenshot" of the current situation,
after monkey patching STDIN the game now is in a constant loop,
(the screen is flickering, hence my assummption)

We can't move forward until we fix this issue.
Although, because the game is rendering correctly I could be tempted to
move forward with further cleaning up.

Except that if I decide to do that I cannot test level transition, which was
also another problem we had and also the reason we are refactoring the code
right now - that reason being that every time we added a new component the game
simply broke. So right now we need to figure out what is going wrong with the game.

No messages yet. Play the game to see messages here.
Seed: 977483248916270 | Rows: 5 | Columns: 5
-----------------------------------

+---+---+---+---+---+
| % |           |   |
+   +   +   +   +   +
| M   M |   |       |
+---+---+   +---+---+
|     M |           |
+   +   +   +---+   +
|   | @ |   |       |
+   +---+---+   +   +
|               |   |
+---+---+---+---+---+

=== MESSAGES ===
No messages yet. Play the game to see messages here.
^C/Users/davidslv/projects/vanilla/lib/vanilla/game.rb:56:in 'Kernel#sleep': Interrupt
        from /Users/davidslv/projects/vanilla/lib/vanilla/game.rb:56:in 'Vanilla::Game#start'
        from ./bin/play.rb:44:in '<main>'