ECS Mermaid documentation
Generated new documentation after the rewrite in PR #50 This has shown that we
are still breaking ECS, this orignated a new document, issue #55 [Understanding
ECS in Game Development](https://github.com/Davidslv/vanilla-private/issues/55)

It is imperative to achieve this as soon as possible, the game currently does
not have many features, and if left unnatended we will struggle to maintain a
good codebase.

We do need to write tooling to prevent developers from making the same mistake
in the future.